Add FlingingRendererClient
Currently, a MojoRenderer is directly used to pass through messages to
and from the FlingingRenderer (FR).

This CL adds the FlingingRendererClient (FRC), which serves almost
exactly the same purpose as a pass-through MojoRenderer. The difference
is that, the FRC implements the mojom::FlingingRendererClientExtension,
which allows the FR to send remote Play/Pause change events directly to
the FRC. This removes the need for the general mojom::RendererClient to
have a specialized OnRemotePlayStateChange().

A follow up CL will remove media::RendererClient::OnRemotePlayState().

Bug: 956677
Change-Id: Ia7484b54849991cfad8a78040d0e47ab156bc3e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1526637
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#654573}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69c2be8680879d2ee74bdbf48b71157a95a8d73e