Fix connection close for streaming preload=metadata resource.
We shouldn't be releasing the connection for streaming resources, even
if they are preload=metadata. It's not possible to resume these
connections, so they must always be held open. We do the same check when
determining if an element can be suspended or not.

BUG=956027
TEST=new test.
R=sandersd

Change-Id: I76610a16d3179a69b76ca52fff50dfdd9f6afb9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1586576
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#654957}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6ccd199e8f1885af1e86540afd6d3747e1b7cea