media: Implement gpu::ImageDecodeAcceleratorWorker for JPEGs on VAAPI.
This CL creates a new class, VaapiJpegDecodeAcceleratorWorker, which
implements the gpu::ImageDecodeAcceleratorWorker interface so that a
gpu::ImageDecodeAcceleratorStub can send JPEG decode requests to the
VAAPI (note that the plumbing between a gpu::ImageDecodeAcceleratorStub
and a VaapiJpegDecodeAcceleratorWorker is left for later).

For now, we only request the decode using VaapiJpegDecoder. In a
follow-up CL, we will extract the decode result and return it.

Bug: 868400
Test: unit tests will be added in a follow-up CL.
Change-Id: I2dafda9c1ca777998e9839fe4a24dd10639831e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1449371
Reviewed-by: Nik Bhagat <nikunjb@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#655859}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 449727e860d5fba0f26e5f04cfd9bbccd9e86a33