Don't cache SkImages wrapping VideoFrame textures
In some cases, to avoid a copy we directly wrap a VideoFrame texture
into a SkImage to use with Canvas. This patch ensures that when we do
so, we don't cache the SkImage if we will release the VideoFrame, as
otherwise we won't synchronize correctly, and might overwrite the
recycled VideoFrame before the Canvas is done accessing it.
When the PaintCanvas passed to Paint() is in software mode, we may
need to read back the VideoFrame texture, in which case Skia will
internally cache the read back pixels, keyed on the SkImage. We still
want to take advantage of that cache, so in this case we force the
copy path so that we can keep the SkImage.

Bug: 958907
Change-Id: I4159646988b1298761cb24209a2d5f72c83f25e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1595294
Commit-Queue: Antoine Labour <piman@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Auto-Submit: Antoine Labour <piman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#658788}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 005d0c604383da8cc7edfcac7f4974ef7730e1d5