media/gpu/linux/GenericDmaBufVFMapper: Map buffer based on |buffer_sizes|
GenericDmabufVideoFrameMapper maps buffers based on dmabuf fds. This is wrong
because the number of buffers can be less than the number of dmabuf fds when
some planes are stored in the same buffer. The correct way is to map buffers with
|buffer_sizes| and compute the address of each plane from each plane's offset.

Bug: None
Test: video_decode_accelerator_tests on kevin and eve
Change-Id: Ife75b487d55a035bf71978f67ac7ab29c55da4a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1609678
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#659833}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5748062ba307ca758df2f36b707abada981fff0c