media/gpu/v4l2svda: fix stateless API specification violation
The stateless API states that resolution of frames should be set on the
OUTPUT queue, and the coded resolution read back from the CAPTURE queue.
However we were setting the resolution on the CAPTURE queue and
expecting the ioctl to update it to the coded resolution.

This works with our current kernel drivers, but is not correct with
respect to the stateless codec API specification. The kernel drivers
have been updated to support the correct behavior, so carry that change
into Chromium as well.

Bug: 948534
Test: Checked that VDA unittest passes on Kevin and Minnie with kernel
driver fixes.

Change-Id: I8f589f81ae1a7d8223235739f0748e58ac25e342
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1553124
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#659876}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91b0391fda87564fe352141cb804d0dfd410d8b6