Add UKM support to media learning.
This CL adds UKM support to DistributionReporter.

As part of this, it adds the UKM source id to ObservationCompletion,
so that it can be sent to the DistributionReporter.

This CL does not enable UKM reporting on any experiment.

Change-Id: I601361ff1b592e053a1a7df922c9381557605e29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1407782
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#660462}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f61550651ddc198e32e8a7e1b39f3138abf3e92