Catch all exceptions thrown by MediaDrm.getPropertyString()
Exceptions are still being thrown when getPropertyString() is called.
They appear to be android.media.ResourceBusyException's. As it's not a
runtime exception and not listed as a possible exception for the method,
catching all exceptions to handle it.

Bug: 965217
Test: MediaDrmBridgeTest media_unittests pass
Change-Id: I5581b5b49c72881133f0efce773d5da6b5dab656
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621426
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: John Rummell <jrummell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#661912}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8391ae5584ee1495c42e937e836489b723d0147c