MSE: Fix undefined shift of negative WebM discard padding
The WebM container spec indicates that a negative discard padding should
correspond to the nanoseconds of decoder output to discard from the
front of the buffer.

This change fixes an undefined shift when a negative discard padding value is
parsed from the stream. It also adds a few notes for potential
pre-existing inconsistency in handling of WebM container discard padding
in the parser and in the frame processor, tracked by bug 969195.

BUG=968952,969195
TEST=fuzzer case no longer repros with this change.

Change-Id: Ibf49edcbe92c7441b44dbe790d0f0fcb3c0f12ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639039
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#665362}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 653d642ffc4f9d59be7806391a24410ca6e4df1e