media/gpu/v4l2svda: stream the OUTPUT queue off when changing resolution
The stateless codec API requires us to set the resolution on the OUTPUT
queue. This however can only be done if the queue is streamed off. Do
this when changing resolution so we can conform to the stateless codec
API.

Bug: 968392
Test: Checked that video_VideoSeek.vp8.switchres passed on Minnie.

Change-Id: I44e50fada7472d159b75711cde3fc0c5661e08e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1635124
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#665819}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7304d68cdb9844b89bd4bd2df7ab3a61c829a958