media/gpu: remove VaapiWrapper::va_surface_format_
This CL is a follow-up to crrev.com/c/1648812, where the lifetime mgmt
of the Vaapi-allocate VASurfaceIDs was moved to the clients. This CL
removes the VaapiWrapper::va_surface_format_, and just keeps the
value on the clients, which anyway have to compute it (and in the case
of the encoder, it's hardcoded).

Test: All unittests and CrOs video playback correct.
Bug: 971891
Change-Id: I81800cec5e62cc2c26fe244e86fb8196a4e072dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1647118
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#667996}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c29cda696d780a263f98750b6c570d848dcca09f