media/gpu/v4l2: Provide the required fds only on DQBUF in V4L2WritableBufferRef::QueueDMABuf()
V4L2WritableBufferRef::QueueDMABuf() checks the number of given fds is the same
as the number of required fds. Because of this check, a caller has to duplicate
their fds in order to extract the expected number of fds. This CL weakens the
check. The function only checks the number of given fds is not less than the
number of required fds, that is, the number of planes of the buffer. A caller
then can pass their fds without duplicating them.

Bug: 941330
Test: video_decode_accelerator tests --use_vd on kevin
Change-Id: Ia32c63c6b6e022ce3cdfc6dfd9d81ccdd813ddc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1654682
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#668360}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1f0a608d4dbf08a26d138467ddf345c85ffc25e