Refactor CdmFile interface to read/write files in the browser process
Instead of returning a base::File object and letting the utility process
do the read/write operations, change the mojo interface so that the operations
are instead done inside the browser process, and pass the contents to/from
the utility process.

This is done to simplify the code and is the first part of a change to make
CdmFile operations use streams instead of base::File.

Bug: 958204
Test: Modified tests pass
Change-Id: I3d01c4a848889ad0f5f3780305a00481b1ae0074
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636438
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#669024}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9fcfaa7623fb97f2f77702f406ebf4ca1d69ecf8