media/gpu/v4l2vd: remove input and output buffer maps
Using input and output buffer maps was a practice of the old
VideoDecodeAccelerators and should not be needed with the new
VideoDecoder interface.

Add a few members and a new constructor to V4L2DecodeSurface so it can
keep track of the input and output buffers as well as the VideoFrame we
will decode into. That way we can get the buffers and frame from the
decode request directly, without the need to look things up and track
their state.

BUG=941330
TEST=VDAtest passes on Minnie.

Change-Id: I21b8c58fb0ee86a670dafa89ba6925968dec77fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1647309
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#669116}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b2196351e163e22cca0b42bcc964b71fdd298b32