media/gpu/vaapi: Use frame pool's NotifyWhenFrameAvailable in VaapiVideoDecoder.
The VaapiVideoDecoder currently attaches at destruction observer to video frames
to get notified when it's returned to the video frame pool. There is a chance
however that the destruction callback will be triggered before the frame is
returned to the pool. This CL uses the newly added NotifyWhenFrameAvailable()
to be notified when the pool has video frames available.

TEST=ran new VD tests on eve

BUG=920058

Change-Id: I52fcb778e93c0c00ecbba495055e0c798cebfb75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1662980
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#670815}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7b5840bc504d5b2f0231cee63c8e1bd06142f15