media/base/VideoFrameLayout: Add is_single_planar and Plane::size, and Remove buffer_sizes
VideoFrameLayout formerly has buffer_sizes that specifies the size of buffers
that is allocated by GBM API. This variable is also used to stand for single and
multi planar that is a concept specifically for V4L2 API. We can also have
(although not always) only one buffer for multi planar format. This double usage
is confusing. Furthermore, it is generally impossible to know the number of
buffers allocated by GBM API.
This adds |size| variable to VideoFrameLayout::Plane and remove buffer_sizes.
The variable always stands for the size of each plane to be mapped. This also
adds |is_single_planar| which should be used in v4l2 code base only in order to
represent single and multi planar with VideoFrameLayout.

Bug: chromium:971943
Test: VDA unittest, VEA unittest, media_unittests
Test: HW video decoder/encoder works with crosvideo.appspot.com and appr.tc

Change-Id: I744ef536857949039a2ccef904d5b09eaffb3f05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1651585
Reviewed-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#671181}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e2d4f4128bd5bd310b5977677de29fa2676eb3ff