media/VideoFrame: use scoped_refptr to store DMABUF fds
Currently, DMABUF-backed VideoFrames directly store their fds as a
vector of ScopedFDs. This has two disadvantages:

1) Wrapping a VideoFrame into another required FD duplication, and
subsequent closing of the duplicated FDs, which results in unneeded
system calls,

2) There is no good way to identify that a wrapped VideoFrame actually
points to the same frame data as its wrappee.

Fix both issues by embedding the ScopedFD vector into a scoped_refptr.
That way the same FDs are used when wrapping a frame (which is safe
since the VideoFrame remains the owner of the FDs), and we can use the
returned vector reference to identify that two VideoFrames actually
point to the same resource as long as one is wrapped into another, which
is a useful feature for some accelerated video decoders.

BUG=chromium:961591
TEST=vda unittest passes on Minnie.
TEST=tast run minnie video.PlayDecodeAccelUsed*

Change-Id: I5ef61692df70021470d4a53c0718deb2cc1a106a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1675283
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#672850}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eadd9d0dc15f5bee756d88ed9873b4632fdea27c