media/gpu: Add a ReleaseSurface() function
This CL adds a pure virtual ReleaseSurface() function to
VaapiImageDecoder that transfers (i.e. gives up) ownership of the VA
surface that contains the result of the last decode to the caller.

The motivation for this change is that we want to be able to export the
surface as a NativePixmap. In order to do that, we need to take
ownership of the surface so that it's not reused by the decoder and the
lifetime of the underlying buffer can be handled externally.

Bug: 868400
Test: jpeg_decode_accelerator_unittest
Change-Id: Ia0949c0c11d11a07d65774222471006830a0b8c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679061
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Gil Dekel <gildekel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673762}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ea7e34c735292756c7d9d71099148821f9162b2