Add IsReadPending to media::DemuxerStream
Plan is to query this method at the time of underflow and categorize
those that do not have pending reads as non-network (machine perf)
underflows. This categorizatioin is approximate: it is possible to
observe pending read's that are not actually holding anything up,
but we expect these to be the minority.

Bug: 912795
Change-Id: I070ecb90ee2577f3d3c50ec8fee0aec5a910e1b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1367105
Auto-Submit: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#674334}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 913c253c9424205fe6404488c51e204e0689dbc2