Remove RendererClient::OnRemotePlayStateChange()
Changes to the Play/Pause state on remote cast devices are propagated
through the general RendererClient and PipelineClient interfaces, via
OnRemotePlayStateChange(). This path is only used a small portion of
the time (when flinging media via the Android-only FlingingRenderer).

This CL adds a path to propagate these events in a much more targeted
way, which scopes the complexity to the classes that actually need to
know about the events.

FlingingRendererClientFactory gives the RendererFactorySelector
a request for a RemotePlayStateChangeCB. This request is fulfilled
in WMPI's constructor. This fixes the chicken/egg problem that arises
from the FRCF being created before WMPI, and minimizes coupling
between classes. The FRCF then passes a copy of the callback to any
renderer it creates.

The end result is that the FlingingRendererClient can directly call
WMPI::OnRemotePlayStateChange(), which is the right place to handle
remote Play/Pause state change events.

Bug: 956677
Change-Id: I89ffaacdb8d142a6317e2ae67924b5fadd373169
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584929
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#674599}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 901808988054ab67dc143f7e43de11dc5980545a