media/gpu: Allocate the pool and converter outside ChromeosVideoDecoderFactory
Originally we created the instances of DmabufVideoFramePool and
VideoFrameConverter inside the factory class. But in this structure
we would have circular dependency.

For example, a client "FooClient" of the factory class needs the
instance of DmabufVideoFramePool "FooFramePool". In recent structure,
the dependency will be:
  FooClient => ChromeosVideoDecoderFactory => FooFramePool.
If both "FooClient" and "FooFramePool" belong to the same build
target, the there is a circular dependency.

In this CL, each client of the factory class needs to allocate the
frame pool and converter instance. Then we won't have circular
dependency.

BUG=b:136716638
TEST=Run video_decode_accelerator_tests on Kevin
TEST=Run Chrome on Kevin and play video

Change-Id: I637354d4d3456a83e8d37776c53c98594b14c0da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680503
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#675508}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0b9a33289b448b90a9c773349854a02f6a2bac87