media/gpu: V4L2SVD: Don't call unnecessary PumpDecodeTask() in EnqueueDecodeTask().
In V4L2SVD, PumpDecodeTask() processes a queue of decoding requests.
The method will process all requests in the queue if possible. So we
don't invoke the method again when we enqueue a new request if there
are already requests in the queue.

BUG=chromium:941330
TEST=video_decode_accelerator tests --use_vd on kevin

Change-Id: I7454ade2c7faa1330f8989e3d0c4ce5b9cbd7513
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690051
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#675522}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a5f9fc1604e79ef27c8a44b6cbdf031af6e40d1