media/gpu/test: Fixed missing frames in video decoder tests in allocate mode.
When running the new video decoder tests in allocate mode, sometimes the video
player will not receive the last frame, making the tests fails. This happens
because the WrapVideoFrame method is only supported in import mode, and a
workaround is used in allocate mode.

To fix this issue logic is added to manually wrap frames in a similar way when
using allocate mode. This is also required to run the video decoder performance
tests on allocate-only platforms.

TEST=./video_decode_accelerator_tests on nyan_big

BUG=957808

Change-Id: Ibf29e9f716beef8aaf1a27904e2220399418c209
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688773
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#676259}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93c4034f8c966fef48b49f4e5f342afa118ed6da