Improve codec checking
In order to determine what codecs are supported by a specified key
system, the current code attempts to create a decoder for each codec to
determine if it is supported. Improve this by simply querying
MediaCodecList if it is available.

This also updates the codec checking to see if VP9 profile 2 is
supported by the device.

Bug: 707127,727918,846120
Test: EncryptedMedia content_browsertests pass
Change-Id: I932b77e0d3d5edb04b0bd8660146875e6c2f16ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1700330
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: John Rummell <jrummell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#677544}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9fd4b6d43b5a8d954db9718133be15f5583c759