Remove checks in ~DecoderBuffer.
The crashes don't appear to be caused by double destruction of the
objects, so removing the check for it. As well, checking that
|side_data| and |side_data_size| are both 0, or both non-zero, causes
crashes when |side_data_size| is corrupted that wouldn't happen
otherwise, so removing the check for it as well.

Bug: 794740
Test: media_unittests pass
Change-Id: I93e3371a2589cbc117ba5fd37eace41de2b06c8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702665
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#677574}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6c29cc708200ac687ae268d72b44a3f05a28629